tests started up

+++ src/art.c
@@ -11,7 +11,7 @@ uint8_t overflow_detected(uint8_t a,
   return 0;
 }
 
-int check_signed_bit(uint8_t number) {
+uint8_t check_signed_bit(uint8_t number) {
   if ((number >> 7)) { // TODO: I don't think I need == 1, should be 0 or 1
     return 1;
   }
@@ -20,7 +20,6 @@ int check_signed_bit(uint8_t number) {
 }
 
 uint8_t check_parity(uint8_t number) {
-  // TODO: resume here, figure out how parity stuff works.
   uint8_t counter = 0;
   for (int x = 0; x < 8; x++) {
     if ((number >> x) & 1) {
@@ -30,7 +29,7 @@ uint8_t check_parity(uint8_t number) {
   return (counter % 2 == 0) ? 1 : 0;
 }
 
-int set_aux_carry(uint8_t number) { // naive implementation
+uint8_t set_aux_carry(uint8_t number) { // naive implementation
   int extract_half_byte = number & 0xf;
 
   if (extract_half_byte > 9) {
@@ -73,8 +72,7 @@ int add_register(CPU *cpu, uint8_t reg_value) {
     cpu->flags.P = 0;
   }
 
-  // aux carry TODO: figure out how this BCD and DAA aux carry stuff works.
-  // Resume here.
+  // aux carry
   if (set_aux_carry(result)) {
     cpu->flags.AC = 1;
   } else {
